+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Add Rails initialization tracking: railtie initializers, `to_prepare` callbacks, and lazy load hooks. ([@ardecvz][])
 
+- Prefix HTTP nodes with `http:`. ([@ardecvz][])
+
 ## 0.2.3 (2026-06-05)
 
 - Add SKILL.md for rails-hyperdrive. ([@palkan][])